Trait move_package::package_hooks::PackageHooks

pub trait PackageHooks {
    fn custom_package_info_fields(&self) -> Vec<String>;
    fn custom_dependency_key(&self) -> Option<String>;
    fn resolve_custom_dependency(
        &self, 
        dep_name: Symbol, 
        info: &CustomDepInfo
    ) -> Result<()>;
}

A trait providing hooks to customize the package system for a particular Move application. An instance of the trait can be registered globally.

Required Methods

fn custom_package_info_fields(&self) -> Vec<String>

Returns custom fields allowed in PackageInfo.

fn custom_dependency_key(&self) -> Option<String>

Returns a custom key for dependencies, if available. This is the string used in dependencies `{ = value, address = addr }.

fn resolve_custom_dependency(
&self,
dep_name: Symbol,
info: &CustomDepInfo
) -> Result<()>

A resolver for custom dependencies in the manifest. This is called to download the dependency from the dependency into the info.local_path location, similar as with git dependencies.