Struct move_package::BuildConfig

pub struct BuildConfig {
    pub dev_mode: bool,
    pub test_mode: bool,
    pub generate_docs: bool,
    pub generate_abis: bool,
    pub install_dir: Option<PathBuf>,
}

Fields
dev_mode: bool

Compile in ‘dev’ mode. The ‘dev-addresses’ and ‘dev-dependencies’ fields will be used if this flag is set. This flag is useful for development of packages that expose named addresses that are not set to a specific value.

test_mode: bool

Compile in ‘test’ mode. The ‘dev-addresses’ and ‘dev-dependencies’ fields will be used along with any code in the ‘test’ directory.

generate_docs: bool

Generate documentation for packages

generate_abis: bool

Generate ABIs for packages

install_dir: Option<PathBuf>

Optional installation directory for this after it has been generated.

Implementations

impl BuildConfig

pub fn compile_package<W: Write>(
self,
path: &Path,
writer: &mut W
) -> Result<CompiledPackage>

pub fn move_model_for_package(
self,
path: &Path,
model_config: ModelConfig
) -> Result<GlobalEnv>

pub fn resolution_graph_for_package(self, path: &Path) -> Result<ResolvedGraph>